refactor: change store and formSlice
