+++ src/components/CardsList.tsx
@@ -1,8 +1,8 @@
-import CardImage from '../CardImage';
-import CardInfo from '../CardInfo';
+import CardImage from './CardImage';
+import CardInfo from './CardInfo';
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { type RootState } from '../form/store';
+import { type RootState } from '../redux/store';
 
 export interface ICardData {
   image: string;
@@ -14,10 +14,10 @@ export interface ICardData {
 }
 
 const CardsList = () => {
-  const cards = useSelector((state: RootState) => state);
-  console.log(cards);
+  const { formDataList } = useSelector((state: RootState) => state.form);
+  console.log(formDataList);
 
-  const RenderCards = cards.formDataList.map((card: ICardData, index: number) => (
+  const RenderCards = formDataList.map((card: ICardData, index: number) => (
     <div className="card" key={index}>
       <CardImage imageCard={card.image} />
       <div className="card_title" role="title">
