+++ src/main.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 
 import { Provider } from 'react-redux';
-import store from './components/form/store';
+import store from './redux/store';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
